refactor(viewer): move sectionWatcher into state directory
