+++ packages/viewer/src/pages/Page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { extractDocPath } from "../state/router.svelte";
   import { getRwContext } from "../lib/context";
-  import { watchPageSection } from "../lib/sectionWatcher.svelte";
+  import { watchPageSection } from "../state/sectionWatcher.svelte";
   import PageContent from "../components/PageContent.svelte";
 
   const { router, page, navigation, liveReload } = getRwContext();
